#89: build: rename header files consistently, fix guards
