+++ examples/serdes_example_3.cc
@@ -42,7 +42,7 @@
 //@HEADER
 */
 
-#include "checkpoint/serdes_headers.h"
+#include <checkpoint/checkpoint.h>
 
 #include <cstdio>
