+++ src/checkpoint/buffer/user_buffer.h
@@ -42,10 +42,10 @@
 //@HEADER
 */
 
-#if !defined INCLUDED_SERDES_USER_BUFFER
-#define INCLUDED_SERDES_USER_BUFFER
+#if !defined INCLUDED_CHECKPOINT_BUFFER_USER_BUFFER_H
+#define INCLUDED_CHECKPOINT_BUFFER_USER_BUFFER_H
 
-#include "checkpoint/serdes_common.h"
+#include "checkpoint/common.h"
 #include "checkpoint/buffer/buffer.h"
 
 namespace serdes {
@@ -71,4 +71,4 @@ private:
 
 } /* end namespace serdes */
 
-#endif /*INCLUDED_SERDES_USER_BUFFER*/
+#endif /*INCLUDED_CHECKPOINT_BUFFER_USER_BUFFER_H*/
