+++ src/checkpoint/common.h
@@ -2,7 +2,7 @@
 //@HEADER
 // *****************************************************************************
 //
-//                            serialize_interface.h
+//                                   common.h
 //                           DARMA Toolkit v. 1.0.0
 //                 DARMA/checkpoint => Serialization Library
 //
@@ -42,50 +42,32 @@
 //@HEADER
 */
 
-#if !defined INCLUDED_SERIALIZE_INTERFACE
-#define INCLUDED_SERIALIZE_INTERFACE
+#if !defined INCLUDED_CHECKPOINT_COMMON_H
+#define INCLUDED_CHECKPOINT_COMMON_H
+
+#define DEBUG_SERDES 0
+
+#if DEBUG_SERDES
+#define debug_serdes(...)            \
+  do {                                          \
+    printf(__VA_ARGS__);             \
+  } while (0);
+#include <cstdio>
+#else
+#define debug_serdes(...)
+#endif
 
 #include <cstdlib>
+#include <cstdint>
 #include <functional>
-#include <memory>
 
-namespace serialization { namespace interface {
+namespace serdes {
 
-using SizeType = size_t;
+using SerialSizeType = size_t;
 using SerialByteType = char;
 
-using BufferCallbackType = std::function<SerialByteType*(SizeType size)>;
+using BufferObtainFnType = std::function<SerialByteType*(SerialSizeType size)>;
 
-struct SerializedInfo {
-  virtual SizeType getSize() const = 0;
-  virtual SerialByteType* getBuffer() const = 0;
-  virtual ~SerializedInfo() { }
-};
+} /* end namespace serdes */
 
-using SerializedInfoPtrType = std::unique_ptr<SerializedInfo>;
-using SerializedReturnType = SerializedInfoPtrType;
-
-template <typename T>
-SerializedReturnType serialize(T& target, BufferCallbackType fn = nullptr);
-
-template <typename T>
-T* deserialize(SerialByteType* buf, SizeType size, T* user_buf = nullptr);
-
-template <typename T>
-void deserializeInPlace(SerialByteType* buf, SizeType size, T* t);
-
-template <typename T>
-T* deserialize(SerializedReturnType&& in);
-
-template <typename T>
-SerializedReturnType serializePartial(T& target, BufferCallbackType fn = nullptr);
-
-template <typename T>
-T* deserializePartial(SerialByteType* buf, SizeType size, T* user_buf = nullptr);
-
-template <typename T>
-std::size_t getSize(T& target);
-
-}} /* end namespace serialization::interface */
-
-#endif /*INCLUDED_SERIALIZE_INTERFACE*/
+#endif /*INCLUDED_CHECKPOINT_COMMON_H*/
