+++ src/checkpoint/container/container_serialize.h
@@ -42,10 +42,10 @@
 //@HEADER
 */
 
-#if !defined INCLUDED_SERDES_CONTAINER_SERIALIZE
-#define INCLUDED_SERDES_CONTAINER_SERIALIZE
+#if !defined INCLUDED_CHECKPOINT_CONTAINER_CONTAINER_SERIALIZE_H
+#define INCLUDED_CHECKPOINT_CONTAINER_CONTAINER_SERIALIZE_H
 
-#include "checkpoint/serdes_common.h"
+#include "checkpoint/common.h"
 #include "checkpoint/serializers/serializers_headers.h"
 
 namespace serdes {
@@ -82,4 +82,4 @@ inline void parserdesContainerElems(Serializer& s, ContainerT& cont) {
 
 } /* end namespace serdes */
 
-#endif /*INCLUDED_SERDES_CONTAINER_SERIALIZE*/
+#endif /*INCLUDED_CHECKPOINT_CONTAINER_CONTAINER_SERIALIZE_H*/
