+++ src/checkpoint/container/view_traverse_ndim.h
@@ -42,10 +42,10 @@
 //@HEADER
 */
 
-#if !defined INCLUDED_CONTAINER_VIEW_TRAVERSE_NDIM_H
-#define INCLUDED_CONTAINER_VIEW_TRAVERSE_NDIM_H
+#if !defined INCLUDED_CHECKPOINT_CONTAINER_VIEW_TRAVERSE_NDIM_H
+#define INCLUDED_CHECKPOINT_CONTAINER_VIEW_TRAVERSE_NDIM_H
 
-#include "checkpoint/serdes_common.h"
+#include "checkpoint/common.h"
 #include "checkpoint/serializers/serializers_headers.h"
 #include "checkpoint/container/view_traits_extract.h"
 
@@ -276,4 +276,4 @@ struct TraverseRecursive {
 
 #endif /*KOKKOS_ENABLED_SERDES*/
 
-#endif /*INCLUDED_SERDES_CONTAINER_VIEW_SERIALIZE_H*/
+#endif /*INCLUDED_CHECKPOINT_CONTAINER_VIEW_TRAVERSE_NDIM_H*/
