+++ src/checkpoint/dispatch/dispatch_common.h
@@ -42,10 +42,10 @@
 //@HEADER
 */
 
-#if !defined INCLUDED_SERDES_DISPATCH_COMMON
-#define INCLUDED_SERDES_DISPATCH_COMMON
+#if !defined INCLUDED_CHECKPOINT_DISPATCH_DISPATCH_COMMON_H
+#define INCLUDED_CHECKPOINT_DISPATCH_DISPATCH_COMMON_H
 
-#include "checkpoint/serdes_common.h"
+#include "checkpoint/common.h"
 #include "checkpoint/buffer/buffer.h"
 #include "checkpoint/dispatch/dispatch_serializer.h"
 #include "checkpoint/dispatch/dispatch_deserializer.h"
@@ -84,4 +84,4 @@ struct DispatchCommon {
 
 } /* end namespace serdes */
 
-#endif /*INCLUDED_SERDES_DISPATCH_COMMON*/
+#endif /*INCLUDED_CHECKPOINT_DISPATCH_DISPATCH_COMMON_H*/
