+++ src/checkpoint/dispatch/dispatch_serializer.h
@@ -42,11 +42,10 @@
 //@HEADER
 */
 
-#if ! defined INCLUDED_SERDES_DISPATCH_SERIALIZER
-#define INCLUDED_SERDES_DISPATCH_SERIALIZER
+#if !defined INCLUDED_CHECKPOINT_DISPATCH_DISPATCH_SERIALIZER_H
+#define INCLUDED_CHECKPOINT_DISPATCH_DISPATCH_SERIALIZER_H
 
-#include "checkpoint/serdes_common.h"
-#include "checkpoint/serdes_all.h"
+#include "checkpoint/common.h"
 #include "checkpoint/traits/serializable_traits.h"
 #include "checkpoint/dispatch/dispatch_serializer_byte.h"
 #include "checkpoint/dispatch/dispatch_serializer_nonbyte.h"
@@ -65,4 +64,4 @@ using SerializerDispatch = SerializerDispatchByte<SerializerT, T>;
 
 } //end namespace serdes
 
-#endif /*INCLUDED_SERDES_DISPATCH_SERIALIZER*/
+#endif /*INCLUDED_CHECKPOINT_DISPATCH_DISPATCH_SERIALIZER_H*/
