+++ src/checkpoint/dispatch/dispatch_serializer_nonbyte.h
@@ -42,11 +42,10 @@
 //@HEADER
 */
 
-#if ! defined INCLUDED_SERDES_DISPATCH_SERIALIZER_NONBYTE
-#define INCLUDED_SERDES_DISPATCH_SERIALIZER_NONBYTE
+#if !defined INCLUDED_CHECKPOINT_DISPATCH_DISPATCH_SERIALIZER_NONBYTE_H
+#define INCLUDED_CHECKPOINT_DISPATCH_DISPATCH_SERIALIZER_NONBYTE_H
 
-#include "checkpoint/serdes_common.h"
-#include "checkpoint/serdes_all.h"
+#include "checkpoint/common.h"
 #include "checkpoint/traits/serializable_traits.h"
 
 #include <type_traits>
@@ -253,4 +252,4 @@ struct SerializerDispatchNonByte {
 
 } //end namespace serdes
 
-#endif /*INCLUDED_SERDES_DISPATCH_SERIALIZER_NONBYTE*/
+#endif /*INCLUDED_CHECKPOINT_DISPATCH_DISPATCH_SERIALIZER_NONBYTE_H*/
