+++ src/checkpoint/serializers/sizer.cc
@@ -42,7 +42,7 @@
 //@HEADER
 */
 
-#include "checkpoint/serdes_common.h"
+#include "checkpoint/common.h"
 #include "checkpoint/serializers/base_serializer.h"
 #include "checkpoint/serializers/sizer.h"
