+++ src/checkpoint/serializers/sizer.h
@@ -42,10 +42,10 @@
 //@HEADER
 */
 
-#if !defined INCLUDED_SERDES_SIZER
-#define INCLUDED_SERDES_SIZER
+#if !defined INCLUDED_CHECKPOINT_SERIALIZERS_SIZER_H
+#define INCLUDED_CHECKPOINT_SERIALIZERS_SIZER_H
 
-#include "checkpoint/serdes_common.h"
+#include "checkpoint/common.h"
 #include "checkpoint/serializers/base_serializer.h"
 
 namespace serdes {
@@ -62,4 +62,4 @@ private:
 
 } /* end namespace serdes */
 
-#endif /*INCLUDED_SERDES_SIZER*/
+#endif /*INCLUDED_CHECKPOINT_SERIALIZERS_SIZER_H*/
