+++ src/checkpoint/traits/serializable_traits.h
@@ -42,10 +42,10 @@
 //@HEADER
 */
 
-#if ! defined INCLUDED_SERDES_SERIALIZABILITY_TRAITS
-#define INCLUDED_SERDES_SERIALIZABILITY_TRAITS
+#if !defined INCLUDED_CHECKPOINT_TRAITS_SERIALIZABLE_TRAITS_H
+#define INCLUDED_CHECKPOINT_TRAITS_SERIALIZABLE_TRAITS_H
 
-#include "checkpoint/serdes_common.h"
+#include "checkpoint/common.h"
 #include "checkpoint/serializers/serializers_headers.h"
 
 #include <cstdint>
@@ -205,4 +205,4 @@ struct SerializableTraits {
 
 #endif  /*HAS_DETECTION_COMPONENT*/
 
-#endif  /*INCLUDED_SERDES_SERIALIZABILITY_TRAITS*/
+#endif /*INCLUDED_CHECKPOINT_TRAITS_SERIALIZABLE_TRAITS_H*/
