+++ tests/unit/test_reconstruct.cc
@@ -44,8 +44,8 @@
 
 #include "test_harness.h"
 
-#include <checkpoint/serdes_headers.h>
-#include <checkpoint/serialization_library_headers.h>
+#include <checkpoint/checkpoint.h>
+#include <checkpoint/checkpoint.h>
 
 #include <gtest/gtest.h>
